Moved kotlin REPL to separate package
